Revised the figure for the models of hiring & output the new `png` figure.
Made a number of changes to the document based on comments by DI.  Added some extra references, used the Choi & Pak paper's definitions of Inter/Multi & Transdisciplinary.  I like the 2+2 model.

Still have a hanging paragraph in front of the Methods section.

A set of changes related to standardizing the case of Ecolog.  Trying to fill in the discussion, left another hanging paragraph.