Edits to `ecolog_paper`:
* Removed the Formatting instructions.

* Fixed table & figure numbers.

* Moved the position of the `weeks` calculation, so that we can use posting numbers higher up in the text, although this is not implemented yet.

Added some code for lemmatization.  This is an alternative to stemming, but is only implemented using the 19th century fiction corpus at present.  It's not really well suited to the text we're looking at.  We could probably make it work better, but at present it results in worse performance than straight stemming.